+++ app/src/main/java/com/example/finedust/setting.kt
@@ -1,0 +1,19 @@
+package com.example.finedust
+
+import android.content.Intent
+import android.support.v7.app.AppCompatActivity
+import android.os.Bundle
+import kotlinx.android.synthetic.main.activity_setting.*
+
+class setting : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_setting)
+
+        button.setOnClickListener {
+            val login_page_move = Intent(this, login_page::class.java)
+            startActivity(login_page_move)
+        }
+    }
+}
