+++ app/src/main/java/com/example/finedust/sign_up_page.kt
@@ -1,0 +1,12 @@
+package com.example.finedust
+
+import android.support.v7.app.AppCompatActivity
+import android.os.Bundle
+
+class sign_up_page : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_sign_up_page)
+    }
+}
