--- app/src/main/java/com/example/misealimi/MainActivity.kt
@@ -1,12 +1,0 @@
-package com.example.misealimi
-
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
