--- app/src/main/java/com/example/misealimi/MyAdapter.kt
@@ -1,4 +1,0 @@
-package com.example.misealimi
-
-class MyAdapter(private val myDataset: Array<String>) {
-}
