+++ app/src/main/res/layout/activity_login_page.xml
@@ -1,0 +1,108 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@color/colorBackground"
+        tools:context=".login_page">
+
+    <ImageView
+            android:layout_centerHorizontal="true"
+            android:layout_width="120dp"
+            android:layout_height="120dp" app:srcCompat="@drawable/goodstate" android:id="@+id/imageView4"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginTop="44dp"
+            app:layout_constraintTop_toTopOf="parent" android:layout_marginStart="8dp"
+            app:layout_constraintHorizontal_bias="0.484" android:layout_marginEnd="8dp"/>
+    <TextView
+            android:layout_centerHorizontal="true"
+            android:textSize="45dp"
+            android:text="환영합니다!"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView10" app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/imageView4" android:layout_marginStart="8dp"
+            android:layout_marginEnd="8dp" android:layout_marginTop="12dp" app:layout_constraintHorizontal_bias="0.6"/>
+    <TextView
+            android:textSize="20dp"
+            android:text="미세먼지 흡입량 맞춤 알리미입니다."
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView11"
+            app:layout_constraintTop_toBottomOf="@+id/textView10" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="8dp" android:layout_marginTop="8dp" app:layout_constraintEnd_toEndOf="parent"
+            android:layout_marginEnd="8dp" app:layout_constraintHorizontal_bias="0.666"/>
+    <TextView
+            android:textSize="20dp"
+            android:text="ID"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView12"
+            app:layout_constraintBottom_toBottomOf="parent"
+            android:layout_marginTop="8dp" app:layout_constraintTop_toBottomOf="@+id/textView11"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="32dp"
+            app:layout_constraintVertical_bias="0.073"/>
+    <TextView
+            android:textSize="20dp"
+            android:text="Password"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView13"
+            app:layout_constraintTop_toBottomOf="@+id/textView12" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="16dp" android:layout_marginTop="40dp"/>
+    <EditText
+            android:hint="ID"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:inputType="date"
+            android:ems="10"
+            android:id="@+id/editText4" app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent" app:layout_constraintHorizontal_bias="0.503"
+            app:layout_constraintVertical_bias="0.052" app:layout_constraintStart_toEndOf="@+id/textView12"
+            android:layout_marginStart="8dp" android:layout_marginTop="8dp"
+            app:layout_constraintTop_toBottomOf="@+id/textView11"/>
+    <EditText
+            android:hint="Password"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:inputType="textPassword"
+            android:ems="10"
+            android:id="@+id/editText5"
+            android:layout_marginStart="24dp"
+            android:layout_marginTop="24dp" app:layout_constraintTop_toBottomOf="@+id/editText4"
+            app:layout_constraintStart_toEndOf="@+id/textView13"/>
+    <Button
+            android:text="login"
+            android:layout_width="168dp"
+            android:layout_height="52dp" android:id="@+id/button6"
+            app:layout_constraintEnd_toEndOf="parent" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="8dp" android:layout_marginEnd="8dp" android:layout_marginTop="28dp"
+            app:layout_constraintTop_toBottomOf="@+id/editText5" app:layout_constraintHorizontal_bias="0.498"/>
+    <ImageButton
+            android:padding="0sp"
+            android:scaleType="centerCrop"
+            android:layout_width="250dp"
+            android:layout_height="40dp" app:srcCompat="@drawable/forgetid" android:id="@+id/imageButton3"
+            app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="16dp"
+            android:layout_marginTop="12dp" app:layout_constraintTop_toBottomOf="@+id/imageButton2"/>
+    <ImageButton
+            android:padding="0sp"
+            android:scaleType="centerCrop"
+            android:layout_width="280dp"
+            android:layout_height="40dp" app:srcCompat="@drawable/forgetpassword"
+            android:id="@+id/imageButton4"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="16dp"
+            app:layout_constraintTop_toBottomOf="@+id/imageButton3" android:layout_marginTop="12dp"/>
+    <ImageButton
+
+            android:padding="0sp"
+            android:scaleType="centerCrop"
+            android:layout_width="220dp"
+            android:layout_height="40dp" app:srcCompat="@drawable/first"
+            android:id="@+id/imageButton2"
+            app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="16dp" android:layout_marginTop="40dp"
+            app:layout_constraintTop_toBottomOf="@+id/button6"/>
+
+</android.support.constraint.ConstraintLayout>
+
