+++ app/src/main/res/layout/activity_sign_up_page.xml
@@ -1,0 +1,129 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@color/colorBackground"
+        tools:context=".sign_up_page">
+
+    <EditText
+            android:layout_width="180dp"
+            android:layout_height="wrap_content"
+            android:inputType="textPassword"
+            android:ems="10"
+            android:id="@+id/editText10"
+            app:layout_constraintStart_toEndOf="@+id/textView20"
+            android:layout_marginTop="20dp" app:layout_constraintTop_toBottomOf="@+id/editText8"
+            android:layout_marginStart="44dp"/>
+    <EditText
+            android:layout_width="180dp"
+            android:layout_height="wrap_content"
+            android:inputType="textPassword"
+            android:ems="10"
+            android:id="@+id/editText9"
+            app:layout_constraintStart_toEndOf="@+id/textView17" android:layout_marginStart="12dp"
+            android:layout_marginTop="20dp" app:layout_constraintTop_toBottomOf="@+id/editText7"/>
+    <EditText
+            android:layout_width="180dp"
+            android:layout_height="wrap_content"
+            android:inputType="textPassword"
+            android:ems="10"
+            android:id="@+id/editText8"
+            app:layout_constraintStart_toEndOf="@+id/textView19" android:layout_marginStart="28dp"
+            android:layout_marginTop="20dp" app:layout_constraintTop_toBottomOf="@+id/editText6"/>
+    <ImageView
+            android:layout_centerHorizontal="true"
+            android:layout_width="120dp"
+            android:layout_height="120dp" app:srcCompat="@drawable/goodstate" android:id="@+id/imageView5"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginTop="44dp"
+            app:layout_constraintTop_toTopOf="parent" android:layout_marginStart="8dp"
+            app:layout_constraintHorizontal_bias="0.487"/>
+    <TextView
+            android:textSize="45dp"
+            android:text="회원가입"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView14"
+            app:layout_constraintTop_toBottomOf="@+id/imageView5" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="124dp" android:layout_marginTop="12dp"/>
+    <TextView
+            android:textSize="20dp"
+            android:text="당신의 정보를 입력해 주세요!"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView15"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="80dp"
+            app:layout_constraintTop_toBottomOf="@+id/textView14" android:layout_marginTop="8dp"/>
+    <TextView
+            android:textSize="25dp"
+            android:text="ID"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView16"
+            app:layout_constraintTop_toBottomOf="@+id/textView15" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="52dp" android:layout_marginTop="36dp"/>
+    <TextView
+            android:textSize="25dp"
+            android:text="Password"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView17"
+            android:layout_marginTop="30dp"
+            app:layout_constraintTop_toBottomOf="@+id/textView16" android:layout_marginStart="8dp"
+            app:layout_constraintStart_toStartOf="parent"/>
+    <TextView
+            android:textSize="25dp"
+            android:text="E-mail"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView18"
+            android:layout_marginTop="30dp"
+            app:layout_constraintTop_toBottomOf="@+id/textView17" android:layout_marginStart="32dp"
+            app:layout_constraintStart_toStartOf="parent"/>
+    <TextView
+            android:textSize="25dp"
+            android:text="몸무게"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView19"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="32dp"
+            app:layout_constraintTop_toBottomOf="@+id/textView18" android:layout_marginTop="30dp"/>
+    <TextView
+            android:textSize="25dp"
+            android:text="나이"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/textView20"
+            app:layout_constraintTop_toBottomOf="@+id/textView19" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="40dp" android:layout_marginTop="30dp"/>
+    <ImageButton
+            android:padding="0sp"
+            android:scaleType="centerCrop"
+            android:layout_width="200dp"
+            android:layout_height="60dp" app:srcCompat="@drawable/signupup" android:id="@+id/imageButton5"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="8dp"
+            android:layout_marginBottom="40dp" app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintHorizontal_bias="0.486"/>
+    <ImageButton
+            android:padding="0sp"
+            android:scaleType="centerCrop"
+            android:layout_width="86dp"
+            android:layout_height="31dp" app:srcCompat="@drawable/doublecheck" android:id="@+id/imageButton8"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toEndOf="@+id/editText7" android:layout_marginStart="5dp"
+            app:layout_constraintHorizontal_bias="0.0"
+            app:layout_constraintBottom_toBottomOf="@+id/editText7" app:layout_constraintTop_toTopOf="@+id/editText7"/>
+    <EditText
+            android:layout_width="180dp"
+            android:layout_height="wrap_content"
+            android:inputType="textPassword"
+            android:ems="10"
+            android:id="@+id/editText6"
+            app:layout_constraintStart_toEndOf="@+id/textView18" android:layout_marginStart="28dp"
+            android:layout_marginTop="20dp" app:layout_constraintTop_toBottomOf="@+id/editText9"/>
+    <EditText
+            android:layout_width="180dp"
+            android:layout_height="44dp"
+            android:inputType="date"
+            android:ems="10"
+            android:id="@+id/editText7"
+            app:layout_constraintStart_toEndOf="@+id/textView16" android:layout_marginStart="56dp"
+            android:layout_marginTop="36dp" app:layout_constraintTop_toBottomOf="@+id/textView15"/>
+</android.support.constraint.ConstraintLayout>
