--- app/src/main/res/layout/gps_timeline.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.RecycleView
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        android:id="@+id/gpsTimelineView"
-        android:scrollbars="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-</android.support.v7.widget.RecycleView>
