+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
-    <string name="app_name">MiseAlimi</string>
+    <string name="app_name">finedust</string>
+    <string name="title_activity_d2">d2</string>
 </resources>
